Working doors

+++ Assets/Scripts/Enemy/EnemyData.cs
@@ -7,10 +7,9 @@ public class EnemyData {
 	public int type; // index of EnemyFactory.enemies.
 	public List<WeaponType> Weaknesses = new List<WeaponType>();
 	public List<WeaponType> Resistences = new List<WeaponType>();
-	public int Life;
+	public float Life;
 
 	public PowerUpData PowerUpData = PowerUpFactory.GetPowerUpNull();
 	public float Scale;
 	public Color ColorOverlay;
-
 }
